refactor(mcp-server): migrate IPC protocol and client to @disclaude/mcp-server (Issue #1042)
Migrates IPC protocol and client code to support the architecture refactoring
goal of separating MCP Server code into its own package.

## Changes

### @disclaude/core
- Add `src/ipc/protocol.ts` - IPC protocol definitions (shared between Primary Node and MCP Server)
- Update exports to include IPC protocol types

### @disclaude/mcp-server
- Add `src/ipc-client/unix-socket-client.ts` - Unix Socket IPC client implementation
- Add `src/ipc-client/index.ts` - IPC client exports
- Update exports to include IPC client

### Backward Compatibility
- `src/ipc/protocol.ts` - Re-exports from @disclaude/core
- `src/ipc/unix-socket-client.ts` - Re-exports from @disclaude/mcp-server

## Verification
- All IPC tests pass (23 tests)
- All MCP tests pass (117 tests)
- Build successful

## Related
- Issue #1042 (MCP Server code migration)
- Issue #1037 (Architecture refactoring parent issue)
- PR #1206 (Previous migration of types and utils)
