+++ packages/core/src/index.ts
@@ -7,6 +7,7 @@
  * - Type definitions (platform, websocket, file)
  * - Constants (deduplication, dialogue, api config)
  * - Utility functions (logger, error-handler, retry)
+ * - IPC Protocol (shared between Primary Node and MCP Server)
  */
 
 // Types
@@ -17,6 +18,9 @@ export * from './constants/index.js';
 
 // Utils
 export * from './utils/index.js';
+
+// IPC Protocol (shared between Primary Node and MCP Server)
+export * from './ipc/index.js';
 
 // Version
 export const CORE_VERSION = '0.0.1';
