+++ packages/core/src/ipc/index.ts
@@ -1,0 +1,15 @@
+/**
+ * IPC exports
+ *
+ * @module core/ipc
+ */
+
+export {
+  DEFAULT_IPC_CONFIG,
+  type IpcConfig,
+  type IpcRequest,
+  type IpcRequestPayloads,
+  type IpcRequestType,
+  type IpcResponse,
+  type IpcResponsePayloads,
+} from './protocol.js';
