+++ packages/mcp-server/src/index.ts
@@ -6,7 +6,7 @@
  * This package contains:
  * - MCP tools types
  * - MCP utilities
- * - IPC client (to be migrated)
+ * - IPC client (for cross-process communication with Primary Node)
  * - MCP resources (to be migrated)
  */
 
@@ -41,6 +41,15 @@ export type {
   ThreadMessageItem,
   GetThreadMessagesResult,
 } from './utils/feishu-api.js';
+
+// IPC Client (Issue #1042: Migrated from src/ipc/)
+export {
+  UnixSocketIpcClient,
+  getIpcClient,
+  resetIpcClient,
+  type IpcAvailabilityStatus,
+  type IpcUnavailableReason,
+} from './ipc-client/index.js';
 
 // Version
 export const MCP_SERVER_VERSION = '0.0.1';
