+++ packages/mcp-server/src/ipc-client/index.ts
@@ -1,0 +1,13 @@
+/**
+ * IPC Client exports
+ *
+ * @module mcp-server/ipc-client
+ */
+
+export {
+  UnixSocketIpcClient,
+  getIpcClient,
+  resetIpcClient,
+  type IpcAvailabilityStatus,
+  type IpcUnavailableReason,
+} from './unix-socket-client.js';
